+++ openvbx.sql
@@ -323,7 +323,8 @@ INSERT INTO settings
 	   ('transcriptions', '1', 1),
 	   ('voice', 'man', 1),
 	   ('voice_language', 'en', 1),
-	   ('numbers_country', 'US', 1);
+	   ('numbers_country', 'US', 1),
+	   ('gravatars', 0, 1);
 
 
 INSERT INTO groups
@@ -331,4 +332,3 @@ INSERT INTO groups
        VALUES
        ('Sales', 1, 1),
        ('Support', 1, 1);
-